even even more refactoring

+++ client/reactive-stocks/src/routes/logout.jsx
@@ -1,27 +1,33 @@
 import { useEffect } from "react";
+import { useNavigate } from "react-router-dom";
 import axios from '../axiosConfig';
+import { checkAuth } from '../utils/auth'; // Import the checkAuth function
 
 export default function Logout() {
     const accessToken = localStorage.getItem('access_token');
+    const navigate = useNavigate();
 
     useEffect(() => {
-        const logout = async () => {
-            try {
-                await axios.post('http://localhost:8000/user_auth/logout/', {
-                    refresh_token: localStorage.getItem('refresh_token')
-                }, { headers: { 'Content-Type': 'application/json', 'Authorization': `Bearer ${accessToken}` } },
-                    { withCredentials: true });
+        async function logout() {
+            const isAuthenticated = await checkAuth(accessToken);
+            if (isAuthenticated) {
+                try {
+                    await axios.post('http://localhost:8000/user_auth/logout/', {
+                        refresh_token: localStorage.getItem('refresh_token')
+                    }, { headers: { 'Content-Type': 'application/json', 'Authorization': `Bearer ${accessToken}` } },
+                        { withCredentials: true });
 
-                localStorage.clear();
-                axios.defaults.headers.common['Authorization'] = null;
-                window.location.href = '/user_auth/login';
-            } catch (e) {
-                console.log('Logout not working', e);
+                    localStorage.clear();
+                    axios.defaults.headers.common['Authorization'] = null;
+                    navigate('/user_auth/login');
+                } catch (e) {
+                    console.log('Logout not working', e);
+                }
             }
-        };
+        }
 
         logout();
-    }, [accessToken]);
+    }, [accessToken, navigate]);
 
     return (
         <div></div>
